azure-functions/requirements.txt
azure-functions==1.20.0
openai>=1.0.0
azure-identity==1.19.0
